refactor(types): unify entity-type registry, remove schedule replay mirror (#571)
khive-pack-schedule's replay-validation path hand-copied
khive-pack-kg's entity-type table (ENTITY_TYPE_DEFS_FOR_REPLAY,
to_snake_case_for_replay, valid_entity_types_for_replay,
resolve_entity_type_for_replay) to work around schedule not being
able to depend on khive-pack-kg. Move the canonical EntityTypeRegistry
(defs, alias resolution, snake_case normalization) into khive-types,
the shared dependency-light crate both packs already depend on, and
delete the hand-copied schedule-side mirror in favor of calling
khive_types::EntityTypeRegistry::global() directly.

khive-pack-kg keeps its entity_type_registry module as a 10-line
re-export so existing import paths keep working. Add
From<khive_types::EntityTypeError> for RuntimeError at the
khive-runtime boundary so existing ?-operator call sites are
unaffected.

Structural only, no behavior change; schedule's parity test against
the KG registry is now backed by the same registry instance rather
than two independently maintained tables.

Deferred: khive-pack-schedule's ScheduleBulkCreateEntryCheck bulk-entry
shape mirror is a different struct (not the entity-type table) and is
left untouched per architect review sign-off.
