+++ crates/khive-pack-schedule/src/handlers.rs
@@ -548,251 +548,13 @@ fn resource_alias_for_replay(normalized: &str) -> bool {
     )
 }
 
-/// Mirror of `khive-pack-kg::entity_type_registry::BUILTIN_DEFS` — one row
-/// per `(kind, canonical type_name, aliases)` triple, used only to reject at
-/// schedule-write-time any `entity_type` that the real KG `create` handler's
-/// `validate_entity_type` (`khive-pack-kg/src/handlers/common.rs`) would
-/// reject at trigger-time replay (round-3 review gap 1). Only the base 8
-/// `khive_types::EntityKind` kinds ever reach this table — see
-/// `validate_entity_type_for_replay` for why `resource` (and any other
-/// pack-owned kind) short-circuits before a subtype is even considered,
-/// mirroring the real handler exactly.
-///
-/// `khive-pack-schedule` does not depend on `khive-pack-kg` in production
-/// (dev-dependency only, for tests), so this hand-copies the definitions
-/// rather than importing `EntityTypeRegistry` directly. Keep in sync with
-/// `BUILTIN_DEFS`:
-/// `gap1_entity_type_exhaustive_parity_with_real_registry` in
-/// `create_validation.rs` compares every entry here against the live
-/// registry (via the dev-dependency, `khive_pack_kg::EntityTypeRegistry::global()`)
-/// and fails if the two tables diverge, so future additions to the real
-/// table that aren't mirrored here fail in CI instead of drifting silently.
-const ENTITY_TYPE_DEFS_FOR_REPLAY: &[(khive_types::EntityKind, &str, &[&str])] = &[
-    // ── Document ──────────────────────────────────────────────────────────
-    (
-        khive_types::EntityKind::Document,
-        "paper",
-        &["preprint", "article"],
-    ),
-    (khive_types::EntityKind::Document, "report", &[]),
-    (khive_types::EntityKind::Document, "blog_post", &["blog"]),
-    (khive_types::EntityKind::Document, "book", &[]),
-    (
-        khive_types::EntityKind::Document,
-        "specification",
-        &["spec"],
-    ),
-    (
-        khive_types::EntityKind::Document,
-        "documentation",
-        &["docs"],
-    ),
-    (khive_types::EntityKind::Document, "thesis", &[]),
-    // ── Concept ───────────────────────────────────────────────────────────
-    (khive_types::EntityKind::Concept, "algorithm", &["algo"]),
-    (
-        khive_types::EntityKind::Concept,
-        "theorem",
-        &["lemma", "proposition", "corollary"],
-    ),
-    (khive_types::EntityKind::Concept, "definition", &["def"]),
-    (
-        khive_types::EntityKind::Concept,
-        "structure",
-        &["inductive", "struct", "class"],
-    ),
-    (khive_types::EntityKind::Concept, "instance", &[]),
-    (khive_types::EntityKind::Concept, "axiom", &[]),
-    (khive_types::EntityKind::Concept, "goal", &["proof_goal"]),
-    (khive_types::EntityKind::Concept, "technique", &[]),
-    (khive_types::EntityKind::Concept, "architecture", &["arch"]),
-    (khive_types::EntityKind::Concept, "model_family", &["model"]),
-    (khive_types::EntityKind::Concept, "theory", &[]),
-    (khive_types::EntityKind::Concept, "research_gap", &["gap"]),
-    (
-        khive_types::EntityKind::Concept,
-        "design_pattern",
-        &["pattern"],
-    ),
-    (
-        khive_types::EntityKind::Concept,
-        "mathematical_operation",
-        &["math_op"],
-    ),
-    (khive_types::EntityKind::Concept, "metric", &[]),
-    (khive_types::EntityKind::Concept, "objective", &["loss"]),
-    // ── Code (ADR-085) ────────────────────────────────────────────────────
-    // "struct"/"class" stay with formal-math "structure" (see above); the
-    // registry does not alias them to code subtypes and neither does this
-    // mirror.
-    (
-        khive_types::EntityKind::Concept,
-        "module",
-        &["mod", "namespace"],
-    ),
-    (
-        khive_types::EntityKind::Concept,
-        "function",
-        &["fn", "func", "method"],
-    ),
-    (
-        khive_types::EntityKind::Concept,
-        "datatype",
-        &["enum", "record", "type_alias"],
-    ),
-    (
-        khive_types::EntityKind::Concept,
-        "interface",
-        &["trait", "protocol"],
-    ),
-    // ── Dataset ───────────────────────────────────────────────────────────
-    (khive_types::EntityKind::Dataset, "benchmark", &[]),
-    (khive_types::EntityKind::Dataset, "corpus", &[]),
-    (
-        khive_types::EntityKind::Dataset,
-        "training_set",
-        &["train_set"],
-    ),
-    (
-        khive_types::EntityKind::Dataset,
-        "evaluation_set",
-        &["eval_set"],
-    ),
-    (khive_types::EntityKind::Dataset, "test_set", &[]),
-    (
-        khive_types::EntityKind::Dataset,
-        "synthetic_dataset",
-        &["synthetic"],
-    ),
-    // ── Project ───────────────────────────────────────────────────────────
-    (khive_types::EntityKind::Project, "library", &["lib"]),
-    (khive_types::EntityKind::Project, "framework", &[]),
-    (khive_types::EntityKind::Project, "tool", &[]),
-    (khive_types::EntityKind::Project, "application", &["app"]),
-    (khive_types::EntityKind::Project, "repository", &["repo"]),
-    // ── Org ───────────────────────────────────────────────────────────────
-    (
-        khive_types::EntityKind::Org,
-        "academic_institution",
-        &["university", "uni"],
-    ),
-    (khive_types::EntityKind::Org, "company", &[]),
-    (khive_types::EntityKind::Org, "research_lab", &["lab"]),
-    (khive_types::EntityKind::Org, "nonprofit", &[]),
-    (
-        khive_types::EntityKind::Org,
-        "government_agency",
-        &["gov_agency"],
-    ),
-    (khive_types::EntityKind::Org, "consortium", &[]),
-    (khive_types::EntityKind::Org, "standards_body", &[]),
-    // ── Artifact ──────────────────────────────────────────────────────────
-    (khive_types::EntityKind::Artifact, "checkpoint", &["ckpt"]),
-    (khive_types::EntityKind::Artifact, "snapshot", &[]),
-    (khive_types::EntityKind::Artifact, "export", &[]),
-    (
-        khive_types::EntityKind::Artifact,
-        "embedding_index",
-        &["embed_index"],
-    ),
-    (khive_types::EntityKind::Artifact, "state_bundle", &[]),
-    (khive_types::EntityKind::Artifact, "profile", &[]),
-    // ── Service ───────────────────────────────────────────────────────────
-    (khive_types::EntityKind::Service, "inference_engine", &[]),
-    (khive_types::EntityKind::Service, "retrieval_engine", &[]),
-    (khive_types::EntityKind::Service, "embedding_engine", &[]),
-    (khive_types::EntityKind::Service, "api", &["endpoint"]),
-    (khive_types::EntityKind::Service, "database", &["db"]),
-    (khive_types::EntityKind::Service, "search_engine", &[]),
-    (khive_types::EntityKind::Service, "mcp_server", &["mcp"]),
-    // Person — no standard subtypes (roles are metadata, not subtypes).
-];
-
-/// Normalize a raw `entity_type` string to canonical snake_case, mirroring
-/// `khive-pack-kg::entity_type_registry::to_snake_case` exactly (ADR-001:106
-/// write-time normalization step that precedes alias resolution): trim ->
-/// lowercase -> runs of separators (space, hyphen, underscore) collapsed to
-/// a single `_` -> leading/trailing `_` stripped.
-fn to_snake_case_for_replay(s: &str) -> String {
-    let mut out = String::with_capacity(s.len());
-    let mut prev_sep = true;
-    for ch in s.chars() {
-        if ch == ' ' || ch == '-' || ch == '_' {
-            if !prev_sep && !out.is_empty() {
-                out.push('_');
-                prev_sep = true;
-            }
-        } else {
-            out.push(ch.to_ascii_lowercase());
-            prev_sep = false;
-        }
-    }
-    if out.ends_with('_') {
-        out.pop();
-    }
-    out
-}
-
-/// Comma-separated list of canonical type names valid for `kind`, mirroring
-/// `EntityTypeRegistry::valid_types_for` for error messages.
-fn valid_entity_types_for_replay(kind: khive_types::EntityKind) -> String {
-    let mut names: Vec<&str> = ENTITY_TYPE_DEFS_FOR_REPLAY
-        .iter()
-        .filter(|(k, _, _)| *k == kind)
-        .map(|(_, type_name, _)| *type_name)
-        .collect();
-    names.sort_unstable();
-    if names.is_empty() {
-        "(none registered)".to_string()
-    } else {
-        names.join(" | ")
-    }
-}
-
-/// Resolve a `(kind, entity_type)` pair against `ENTITY_TYPE_DEFS_FOR_REPLAY`,
-/// mirroring `EntityTypeRegistry::resolve` exactly: kind-qualified lookup
-/// first (a name or alias registered under this specific kind), then a bare
-/// lookup across all kinds — if the name/alias exists under a *different*
-/// kind, the cross-kind rejection error names the correct kind (matching
-/// the real registry's ambiguity-safe bare-name path); otherwise the type is
-/// entirely unknown.
-fn resolve_entity_type_for_replay(
-    kind: khive_types::EntityKind,
-    raw_type: &str,
-) -> Result<Option<String>, RuntimeError> {
-    let normalized = to_snake_case_for_replay(raw_type.trim());
-
-    for (def_kind, type_name, aliases) in ENTITY_TYPE_DEFS_FOR_REPLAY.iter().copied() {
-        if def_kind == kind
-            && (type_name == normalized.as_str() || aliases.contains(&normalized.as_str()))
-        {
-            return Ok(Some(type_name.to_string()));
-        }
-    }
-
-    for (def_kind, type_name, aliases) in ENTITY_TYPE_DEFS_FOR_REPLAY.iter().copied() {
-        if type_name == normalized.as_str() || aliases.contains(&normalized.as_str()) {
-            return Err(RuntimeError::InvalidInput(format!(
-                "entity_type {raw_type:?} belongs to {:?}, not {:?}; valid types for {:?}: {}",
-                def_kind.name(),
-                kind.name(),
-                kind.name(),
-                valid_entity_types_for_replay(kind),
-            )));
-        }
-    }
-
-    Err(RuntimeError::InvalidInput(format!(
-        "unknown entity_type {raw_type:?} for {:?}; valid: {}",
-        kind.name(),
-        valid_entity_types_for_replay(kind),
-    )))
-}
-
 /// Validate an `entity_type` value the same way
 /// `khive-pack-kg::handlers::common::validate_entity_type` does: parse
 /// `canonical_kind_name` into the base `khive_types::EntityKind` first, then
-/// resolve the subtype against `ENTITY_TYPE_DEFS_FOR_REPLAY`.
+/// resolve the subtype against the canonical `khive_types::EntityTypeRegistry`
+/// (#571). KG create validation and schedule replay validation now resolve
+/// against the exact same registry instance instead of two hand-maintained
+/// copies, so this can no longer drift from the real handler's vocabulary.
 ///
 /// The kind-parse step is exactly what makes a pack-owned kind like
 /// `"resource"` reject *any* non-`None` `entity_type` outright: `resource`
@@ -815,7 +577,10 @@ fn validate_entity_type_for_replay(
         .map_err(|_| {
             RuntimeError::InvalidInput(format!("unknown entity kind {canonical_kind_name:?}"))
         })?;
-    resolve_entity_type_for_replay(kind, raw)
+    khive_types::EntityTypeRegistry::global()
+        .resolve(kind, Some(raw))
+        .map(|resolved| resolved.entity_type)
+        .map_err(RuntimeError::from)
 }
 
 /// Reconcile a granular `kind`'s resolved `specific` value against a legacy
